+++ services/dataprotection/mgmt/2021-01-01/dataprotection/dataprotectionapi/interfaces.go
@@ -12,26 +12,17 @@ import (
 	"github.com/Azure/go-autorest/autorest"
 )
 
-// BaseClientAPI contains the set of methods on the BaseClient type.
-type BaseClientAPI interface {
-	CheckFeatureSupport(ctx context.Context, location string, parameters dataprotection.BasicFeatureValidationRequestBase) (result dataprotection.FeatureValidationResponseBaseModel, err error)
-	GetOperationResultPatch(ctx context.Context, vaultName string, resourceGroupName string, operationID string) (result dataprotection.BackupVaultResource, err error)
-	GetOperationStatus(ctx context.Context, location string, operationID string) (result dataprotection.OperationResource, err error)
-}
-
-var _ BaseClientAPI = (*dataprotection.BaseClient)(nil)
-
 // BackupVaultsClientAPI contains the set of methods on the BackupVaultsClient type.
 type BackupVaultsClientAPI interface {
 	CheckNameAvailability(ctx context.Context, resourceGroupName string, location string, parameters dataprotection.CheckNameAvailabilityRequest) (result dataprotection.CheckNameAvailabilityResult, err error)
 	CreateOrUpdate(ctx context.Context, vaultName string, resourceGroupName string, parameters dataprotection.BackupVaultResource) (result dataprotection.BackupVaultsCreateOrUpdateFuture, err error)
 	Delete(ctx context.Context, vaultName string, resourceGroupName string) (result autorest.Response, err error)
 	Get(ctx context.Context, vaultName string, resourceGroupName string) (result dataprotection.BackupVaultResource, err error)
-	GetResourcesInResourceGroup(ctx context.Context, resourceGroupName string) (result dataprotection.BackupVaultResourceListPage, err error)
-	GetResourcesInResourceGroupComplete(ctx context.Context, resourceGroupName string) (result dataprotection.BackupVaultResourceListIterator, err error)
-	GetResourcesInSubscription(ctx context.Context) (result dataprotection.BackupVaultResourceListPage, err error)
-	GetResourcesInSubscriptionComplete(ctx context.Context) (result dataprotection.BackupVaultResourceListIterator, err error)
-	Patch(ctx context.Context, vaultName string, resourceGroupName string, parameters dataprotection.PatchResourceRequestInput) (result dataprotection.BackupVaultsPatchFuture, err error)
+	GetInResourceGroup(ctx context.Context, resourceGroupName string) (result dataprotection.BackupVaultResourceListPage, err error)
+	GetInResourceGroupComplete(ctx context.Context, resourceGroupName string) (result dataprotection.BackupVaultResourceListIterator, err error)
+	GetInSubscription(ctx context.Context) (result dataprotection.BackupVaultResourceListPage, err error)
+	GetInSubscriptionComplete(ctx context.Context) (result dataprotection.BackupVaultResourceListIterator, err error)
+	Update(ctx context.Context, vaultName string, resourceGroupName string, parameters dataprotection.PatchResourceRequestInput) (result dataprotection.BackupVaultsUpdateFuture, err error)
 }
 
 var _ BackupVaultsClientAPI = (*dataprotection.BackupVaultsClient)(nil)
@@ -42,6 +33,27 @@ type OperationResultClientAPI interface {
 }
 
 var _ OperationResultClientAPI = (*dataprotection.OperationResultClient)(nil)
+
+// OperationStatusClientAPI contains the set of methods on the OperationStatusClient type.
+type OperationStatusClientAPI interface {
+	Get(ctx context.Context, location string, operationID string) (result dataprotection.OperationResource, err error)
+}
+
+var _ OperationStatusClientAPI = (*dataprotection.OperationStatusClient)(nil)
+
+// BackupVaultOperationResultsClientAPI contains the set of methods on the BackupVaultOperationResultsClient type.
+type BackupVaultOperationResultsClientAPI interface {
+	Get(ctx context.Context, vaultName string, resourceGroupName string, operationID string) (result dataprotection.BackupVaultResource, err error)
+}
+
+var _ BackupVaultOperationResultsClientAPI = (*dataprotection.BackupVaultOperationResultsClient)(nil)
+
+// ClientAPI contains the set of methods on the Client type.
+type ClientAPI interface {
+	CheckFeatureSupport(ctx context.Context, location string, parameters dataprotection.BasicFeatureValidationRequestBase) (result dataprotection.FeatureValidationResponseBaseModel, err error)
+}
+
+var _ ClientAPI = (*dataprotection.Client)(nil)
 
 // OperationsClientAPI contains the set of methods on the OperationsClient type.
 type OperationsClientAPI interface {
@@ -73,47 +85,35 @@ type BackupInstancesClientAPI interface {
 	TriggerRehydrate(ctx context.Context, resourceGroupName string, vaultName string, parameters dataprotection.AzureBackupRehydrationRequest, backupInstanceName string) (result dataprotection.BackupInstancesTriggerRehydrateFuture, err error)
 	TriggerRestore(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, parameters dataprotection.BasicAzureBackupRestoreRequest) (result dataprotection.BackupInstancesTriggerRestoreFuture, err error)
 	ValidateForBackup(ctx context.Context, vaultName string, resourceGroupName string, parameters dataprotection.ValidateForBackupRequest) (result dataprotection.BackupInstancesValidateForBackupFuture, err error)
-	ValidateRestore(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, parameters dataprotection.ValidateRestoreRequestObject) (result dataprotection.BackupInstancesValidateRestoreFuture, err error)
+	ValidateForRestore(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, parameters dataprotection.ValidateRestoreRequestObject) (result dataprotection.BackupInstancesValidateForRestoreFuture, err error)
 }
 
 var _ BackupInstancesClientAPI = (*dataprotection.BackupInstancesClient)(nil)
 
 // RecoveryPointsClientAPI contains the set of methods on the RecoveryPointsClient type.
 type RecoveryPointsClientAPI interface {
-	GetList(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, filter string, skipToken string) (result dataprotection.AzureBackupRecoveryPointResourceListPage, err error)
-	GetListComplete(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, filter string, skipToken string) (result dataprotection.AzureBackupRecoveryPointResourceListIterator, err error)
+	Get(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, recoveryPointID string) (result dataprotection.AzureBackupRecoveryPointResource, err error)
+	List(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, filter string, skipToken string) (result dataprotection.AzureBackupRecoveryPointResourceListPage, err error)
+	ListComplete(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, filter string, skipToken string) (result dataprotection.AzureBackupRecoveryPointResourceListIterator, err error)
 }
 
 var _ RecoveryPointsClientAPI = (*dataprotection.RecoveryPointsClient)(nil)
 
-// RecoveryPointClientAPI contains the set of methods on the RecoveryPointClient type.
-type RecoveryPointClientAPI interface {
-	Get(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, recoveryPointID string) (result dataprotection.AzureBackupRecoveryPointResource, err error)
-}
-
-var _ RecoveryPointClientAPI = (*dataprotection.RecoveryPointClient)(nil)
-
 // JobsClientAPI contains the set of methods on the JobsClient type.
 type JobsClientAPI interface {
+	Get(ctx context.Context, resourceGroupName string, vaultName string, jobID string) (result dataprotection.AzureBackupJobResource, err error)
 	List(ctx context.Context, resourceGroupName string, vaultName string) (result dataprotection.AzureBackupJobResourceListPage, err error)
 	ListComplete(ctx context.Context, resourceGroupName string, vaultName string) (result dataprotection.AzureBackupJobResourceListIterator, err error)
 }
 
 var _ JobsClientAPI = (*dataprotection.JobsClient)(nil)
 
-// FindRestorableTimeRangesClientAPI contains the set of methods on the FindRestorableTimeRangesClient type.
-type FindRestorableTimeRangesClientAPI interface {
-	Post(ctx context.Context, vaultName string, resourceGroupName string, backupInstances string, parameters dataprotection.AzureBackupFindRestorableTimeRangesRequest) (result dataprotection.AzureBackupFindRestorableTimeRangesResponseResource, err error)
+// RestorableTimeRangesClientAPI contains the set of methods on the RestorableTimeRangesClient type.
+type RestorableTimeRangesClientAPI interface {
+	Find(ctx context.Context, vaultName string, resourceGroupName string, backupInstanceName string, parameters dataprotection.AzureBackupFindRestorableTimeRangesRequest) (result dataprotection.AzureBackupFindRestorableTimeRangesResponseResource, err error)
 }
 
-var _ FindRestorableTimeRangesClientAPI = (*dataprotection.FindRestorableTimeRangesClient)(nil)
-
-// JobClientAPI contains the set of methods on the JobClient type.
-type JobClientAPI interface {
-	Get(ctx context.Context, resourceGroupName string, vaultName string, jobID string) (result dataprotection.AzureBackupJobResource, err error)
-}
-
-var _ JobClientAPI = (*dataprotection.JobClient)(nil)
+var _ RestorableTimeRangesClientAPI = (*dataprotection.RestorableTimeRangesClient)(nil)
 
 // ExportJobsClientAPI contains the set of methods on the ExportJobsClient type.
 type ExportJobsClientAPI interface {